@@ -1,29 +1,29 @@
 import EventEmitter from 'event-emitter';
 import keydrown from 'keydrown';
-import keys from './keys';
+import KEYS from './keys';
 
 const emitter = new EventEmitter();
 
-for (let key of Object.keys(keys)) {
-    keydrown[key].down(() => {
-        emitter.emit(key, key);
-    })
+for (const key of Object.keys(KEYS)) {
+  keydrown[key].down(() => {
+    emitter.emit(key, key);
+  });
 }
 
 class KeyEventEmitter {
-    on(keys, onPress) {
-        let keyArray = [].concat(keys);
-        for (let key of keys) {
-            emitter.on(key, onPress);
-        }
+  on(keys, onPress) {
+    const keyArray = [].concat(keys);
+    for (const key of keyArray) {
+      emitter.on(key, onPress);
     }
+  }
 
-    off(keys, onPress) {
-        let keyArray = [].concat(keys);
-        for (let key of keys) {
-            emitter.off(key, onPress);
-        }
+  off(keys, onPress) {
+    const keyArray = [].concat(keys);
+    for (const key of keyArray) {
+      emitter.off(key, onPress);
     }
+  }
 }
 
 export default new KeyEventEmitter();
