@@ -1,13 +1,13 @@
 const KEYS = {
-    A: 'A',
-    D: 'D',
-    I: 'I',
-    J: 'J',
-    K: 'K',
-    L: 'L',
-    S: 'S',
-    W: 'W',
-    ENTER: 'ENTER'
+  A: 'A',
+  D: 'D',
+  I: 'I',
+  J: 'J',
+  K: 'K',
+  L: 'L',
+  S: 'S',
+  W: 'W',
+  ENTER: 'ENTER',
 };
 
 export default KEYS;
