@@ -1,20 +1,19 @@
 import ACTIONS from '../actions';
-import {
-    nextPosition
-} from './Movement.js';
+import { nextPosition } from './Movement.js';
+
 const Three = require('three');
 
-export const CharacterInitialState = {
-    characterPosition: new Three.Vector3(0, 0, 0)
+export const characterInitialState = {
+  characterPosition: new Three.Vector3(0, 0, 0),
 };
 
-export const CharacterReducer = (state = CharacterInitialState, action) => {
-    switch (action.type) {
-        case ACTIONS.CHARACTER.WALK:
-            return {
-                characterPosition: nextPosition(state.characterPosition, action),
-            };
-        default:
-            return state;
-    }
+export const characterReducer = (state = characterInitialState, action) => {
+  switch (action.type) {
+    case ACTIONS.CHARACTER.WALK:
+      return {
+        characterPosition: nextPosition(state.characterPosition, action),
+      };
+    default:
+      return state;
+  }
 };
